// lib/api/generated/model/requirementApprovalRequestCreate.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 要件定義承認申請リクエストで受け取るschema。
 */
export interface RequirementApprovalRequestCreate {
  target_type: string;
  target_id: number;
  approver_id: number;
  comment?: string | null;
}