// lib/api/generated/model/requirementDetailCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */
import type { RequirementDetailCreateDetailJson } from "./requirementDetailCreateDetailJson";

/**
 * 要件詳細作成リクエストで受け取るschema。
 */
export interface RequirementDetailCreate {
  detail_type: string;
  detail_json: RequirementDetailCreateDetailJson;
}